Make userChoice case-insensitive
Replaced the old way to make userChoice
case-insensitive as it was inefficient.
Now the prompt message the user inputs will
always lowercase, as well as the array elements
in the for loop so that they actually match to
"userChoice". This implies that even if the user
types "SciSsors", it will be lowercased and turn
to "scissors", which will be compared to the array
element "scissors", and ultimately if the
comparison is true, will return "Scissors"