Fix prompt error when null
If the user canceled the prompt or didn't type
anything, an error would be thrown as it resulted
in null variable. A new variable "userChoiceToLC"
has been added, which stores "userChoice" and
lowercase it, but it also makes sure that the
prompt isn't null. Instead, if user doesn't
type anything, the prompt will be an empty
string