Edit case-insensitivity and stop game if no input
Made changes to the case-insensitive prompt.
Now, it will only lowercase "userChoiceToLC" if
the prompt in "userChoice" is a string. If the
prompt is empty, and so is null, it will stay this
way. The error regarding lowercase was due to the
fact that the lowercase method doesn't work on a
null variable. Also, the "isFound" variable is now
declared within the global scope, as it is also
used in the "playRound" function; this function
now checks if the variable "isFound" is true
after running "getUserChoice", which always sets
"isFound" to false unless the data input by the
user matches one of the game choices. Before,
the game would keep playing despite "userChoice"
value not being found, i.e. when user doesn't type
anything or if the choice written is incorrect.
Now, it will stop