Add playGame function & roundCount
New function playGame added, within it
five rounds will play. This function's job is
to call the playRounds function, to reassign the
value of the variables "humanSelection" and
"computerSelection" after every round. It also
displays a message about who won the game at the
end. Also added the "roundCount" global scope
variable to keep track of the actual rounds played
during the game. Every time a round is played and
the user actually inputs a choice that matches,
the "roundCount" increases by one