Implement user interface
The game now plays with a UI that allows the
player to make his choices via three buttons
instead of a prompt. The results are showed
with text on screen as well as the score which
updates. The player who reaches 5 points first
is declared the winner and message is shown,
terminating the game. The "userChoice" function
has been removed as it now no longer serves a
purpose and there have been several changes
including css styling for the score